fix: 팀원 코드 병합 후 버그 수정
다른 팀원의 코드를 병합하고 테스트를 진행하던 중 발견된 버그를 수정하였습니다.
API 호출 로직에 문제가 있어 발생한 것으로 확인되었으며, 이를 수정하여 정상적으로 동작하도록 하였습니다.
