+++ .env.example
@@ -1,5 +1,15 @@
-DATABASE_URL=""
+DATABASE_URL=
 PORT=
 
+EMAIL_SERVICE=
 EMAIL_ADDRESS=
 EMAIL_PASSWORD=
+
+NAVER_CLIENT_ID=
+NAVER_SECRET=
+NAVER_CALLBACK=
+
+ACCESS_TOKEN_SECRET=
+REFRESH_TOKEN_SECRET=
+ACCESS_TOKEN_EXPIRES_IN=
+REFRESH_TOKEN_EXPIRES_IN=
