+++ src/constants/env.constant.js
@@ -9,5 +9,7 @@ export const ENV_KEY = {
   NAVER_SECRET: process.env.NAVER_SECRET,
   NAVER_CALLBACK: process.env.NAVER_CALLBACK,
   ACCESS_TOKEN_SECRET: process.env.ACCESS_TOKEN_SECRET,
-  REFRESH_TOKEN_SECRET: process.env.REFRESH_TOKEN_SECRET
+  REFRESH_TOKEN_SECRET: process.env.REFRESH_TOKEN_SECRET,
+  ACCESS_TOKEN_EXPIRES_IN: process.env.ACCESS_TOKEN_EXPIRES_IN,
+  REFRESH_TOKEN_EXPIRES_IN: process.env.REFRESH_TOKEN_EXPIRES_IN
 }
