+++ src/middlwarmies/require-access-token.middleware.js
@@ -1,76 +1,76 @@
 import jwt from 'jsonwebtoken';
 import { HTTP_STATUS } from '../constants/http-status.constant.js';
 import { MESSAGES } from '../constants/message.constant.js';
-import { ACCESS_TOKEN_SECRET } from '../constants/env.constant.js';
-import { prisma } from '../utils/prisma.util.js';
+import { ENV_KEY } from '../constants/env.constant.js';
+import { prisma } from '../utils/prisma.utils.js';
 
 export const requireAccessToken = async (req, res, next) => {
-    try {
-        // 인증 정보 파싱
-        const authorization = req.headers.authorization;
+  try {
+    // 인증 정보 파싱
+    const authorization = req.headers.authorization;
 
-        // Authorization이 없는 경우
-        if (!authorization) {
-            return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                status: HTTP_STATUS.UNAUTHORIZED,
-                message: MESSAGES.AUTH.COMMON.JWT.NO_TOKEN,
-            });
-        }
-        // JWT 표준 인증 형태와 일치하지 않는 경우
-        const [type, accessToken] = authorization.split(' ');
-
-        if (type !== 'Bearer') {
-            return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                status: HTTP_STATUS.UNAUTHORIZED,
-                message: MESSAGES.AUTH.COMMON.JWT.NOT_SUPPORTED_TYPE,
-            });
-        }
-
-        // AccessToken이 없는 경우
-        if (!accessToken) {
-            return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                status: HTTP_STATUS.UNAUTHORIZED,
-                message: MESSAGES.AUTH.COMMON.JWT.NO_TOKEN,
-            });
-        }
-
-        let payload;
-        try {
-            payload = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
-        } catch (error) {
-            // AccessToken의 유효기한이 지난 경우
-            if (error.name === 'TokenExpiredError') {
-                return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                    status: HTTP_STATUS.UNAUTHORIZED,
-                    message: MESSAGES.AUTH.COMMON.JWT.EXPIRED,
-                });
-            }
-            // 그 밖의 AccessToken 검증에 실패한 경우
-            else {
-                return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                    status: HTTP_STATUS.UNAUTHORIZED,
-                    message: MESSAGES.AUTH.COMMON.JWT.INVALID,
-                });
-            }
-        }
-
-        // Payload에 담긴 사용자 ID와 일치하는 사용자가 없는 경우
-        const { id } = payload;
-        const user = await prisma.user.findUnique({
-            where: { id },
-            omit: { password: true },
-        });
-
-        if (!user) {
-            return res.status(HTTP_STATUS.UNAUTHORIZED).json({
-                status: HTTP_STATUS.UNAUTHORIZED,
-                message: MESSAGES.AUTH.COMMON.JWT.NO_USER,
-            });
-        }
-
-        req.user = user;
-        next();
-    } catch (error) {
-        next(error);
+    // Authorization이 없는 경우
+    if (!authorization) {
+      return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+        status: HTTP_STATUS.UNAUTHORIZED,
+        message: MESSAGES.AUTH.COMMON.JWT.NO_TOKEN,
+      });
     }
+    // JWT 표준 인증 형태와 일치하지 않는 경우
+    const [type, accessToken] = authorization.split(' ');
+
+    if (type !== 'Bearer') {
+      return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+        status: HTTP_STATUS.UNAUTHORIZED,
+        message: MESSAGES.AUTH.COMMON.JWT.NOT_SUPPORTED_TYPE,
+      });
+    }
+
+    // AccessToken이 없는 경우
+    if (!accessToken) {
+      return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+        status: HTTP_STATUS.UNAUTHORIZED,
+        message: MESSAGES.AUTH.COMMON.JWT.NO_TOKEN,
+      });
+    }
+
+    let payload;
+    try {
+      payload = jwt.verify(accessToken, ENV_KEY.ACCESS_TOKEN_SECRET);
+    } catch (error) {
+      // AccessToken의 유효기한이 지난 경우
+      if (error.name === 'TokenExpiredError') {
+        return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+          status: HTTP_STATUS.UNAUTHORIZED,
+          message: MESSAGES.AUTH.COMMON.JWT.EXPIRED,
+        });
+      }
+      // 그 밖의 AccessToken 검증에 실패한 경우
+      else {
+        return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+          status: HTTP_STATUS.UNAUTHORIZED,
+          message: MESSAGES.AUTH.COMMON.JWT.INVALID,
+        });
+      }
+    }
+
+    // Payload에 담긴 사용자 ID와 일치하는 사용자가 없는 경우
+    const { id } = payload;
+    console.log(id);
+    const user = await prisma.user.findUnique({
+      where: { userId: id },
+      // omit: { password: true },
+    });
+
+    if (!user) {
+      return res.status(HTTP_STATUS.UNAUTHORIZED).json({
+        status: HTTP_STATUS.UNAUTHORIZED,
+        message: MESSAGES.AUTH.COMMON.JWT.NO_USER,
+      });
+    }
+    req.user = user;
+    next();
+  } catch (error) {
+    next(error);
+  }
 };
