+++ src/middlwarmies/validation/sign-in-validator.middleware.js
@@ -1,5 +1,5 @@
-import Joi from 'Joi';
-import { MESSAGES } from '../constants/message.constant.js';
+import Joi from 'joi';
+import { MESSAGES } from '../../constants/message.constant.js';
 const schema = Joi.object({
     email: Joi.string().email().required().messages({
         'any.required': MESSAGES.AUTH.COMMON.EMAIL.REQUIRED,
@@ -10,7 +10,6 @@ const schema = Joi.object({
     }),
 
 })
-
 
 export const signinValidator = async (req, res, next) => {
     try {
